Moved resource classes to common for sharing with controller
